requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
Click==7.0
Flask==1.1.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mockredispy==2.9.3
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.2.2
redis==3.3.11
six==1.12.0
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
